feat: reactStrictMode to true + call add-row to last button action + Moving 'voitures 'in the current simulations
