+++ src/app/(layout-with-navigation)/(simulation)/fin/page.tsx
@@ -7,39 +7,8 @@ export default function FinPage() {
   const [_data, setData] = useState(null);
   const [canEndSurvey, setCanEndSurvey] = useState<boolean>(false);
 
-  useEffect(() => {
-    const localStorageValue = localStorage.getItem('nosgestesempreinte::v1')
-    let value = null
-    if (localStorageValue) {
-      const JSONValue = JSON.parse(localStorageValue)
-      //TODO: Pour l'instant on prend la dernière mais à voir pour la suite
-      JSONValue.simulation = JSONValue.simulations.at(-1)
-      delete JSONValue.simulations
-      value = JSON.stringify(JSONValue)
-    }
-
-    fetch('/api/add-row', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ data: value }),
-    })
-        .then((res) => res.json())
-        .then((data) => {
-          console.log('Réponse du serveur:', data);
-          setData(data.message);
-          setCanEndSurvey(true);
-        })
-        .catch((error) => {
-          console.error('Erreur lors de l\'envoi de la requête:', error);
-        });
-  }, []);
-
   const goToOpinionWay = () => {
     if (!canEndSurvey) return;
-
-    window.location.replace('/');
   }
 
   return (
